Moving bigquery v2 under bigquery_v2
