+++ metadata-ingestion/src/datahub/ingestion/source/bigquery_v2/lineage.py
@@ -8,16 +8,16 @@ from google.cloud.logging_v2.client import Client as GCPLoggingClient
 from ratelimiter import RateLimiter
 
 from datahub.emitter import mce_builder
-from datahub.ingestion.source.bigquery.bigquery_audit import (
+from datahub.ingestion.source.bigquery_v2.bigquery_audit import (
     AuditLogEntry,
     BigQueryAuditMetadata,
     BigqueryTableIdentifier,
     BigQueryTableRef,
     QueryEvent,
 )
-from datahub.ingestion.source.bigquery.bigquery_config import BigQueryV2Config
-from datahub.ingestion.source.bigquery.bigquery_report import BigQueryV2Report
-from datahub.ingestion.source.bigquery.common import (
+from datahub.ingestion.source.bigquery_v2.bigquery_config import BigQueryV2Config
+from datahub.ingestion.source.bigquery_v2.bigquery_report import BigQueryV2Report
+from datahub.ingestion.source.bigquery_v2.common import (
     BQ_DATE_SHARD_FORMAT,
     BQ_DATETIME_FORMAT,
     _make_gcp_logging_client,
